// enums/Tips.enum.ts
export enum TipsEnrollmentStatus {
    Enroll = 'enroll',
    Enrolled = 'enrolled',
    Incomplete = 'incomplete',
    Completed = 'completed',
}

export enum TipsEnrollementTypes {
    Current = 'current',
    Completed = 'completed'
}

export enum TipsAttachmentGeneralTypes {
    Uploadable = 'uploadable',
    Embeddable = 'embeddable'
}

export enum TipsAttachmentContentTypes {
    Image = 'image',
    Audio = 'audio',
    Video = 'video',
    Link = 'link',
    Pdf = 'pdf',
    Doc = 'doc',
    Other = 'others'
}

export enum TipsAttachmentVendorTypes {
    Soundcloud = 'soundcloud',
    Youtube = 'youtube',
    Vimeo = 'vimeo'
}